requirements.txt
customtkinter
python-whois
requests